requirements.txt
pyannote.audio==3.3.1
scikit-learn==1.5.1
librosa==0.10.2.post1
